requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7
python-dotenv==1.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0
pytest==7.4.2
black==23.7.0
flake8==6.1.0